extract modals from MaterialCard into their own components
